+++ lib/Expression/Evaluator.php
@@ -6,21 +6,8 @@ use PhpBench\Expression\Ast\Node;
 use PhpBench\Expression\Exception\EvaluationError;
 use PhpBench\Expression\Exception\ExpressionError;
 
-final class Evaluator
+interface Evaluator
 {
-    /**
-     * @var NodeEvaluator<Node>[]
-     */
-    private $evaluators;
-
-    /**
-     * @param NodeEvaluator<Node>[] $evaluators
-     */
-    public function __construct(array $evaluators)
-    {
-        $this->evaluators = $evaluators;
-    }
-
     /**
      * @template T of Node
      *
@@ -29,35 +16,10 @@ final class Evaluator
      *
      * @return T
      */
-    public function evaluateType(Node $node, string $expectedType, array $params): Node
-    {
-        $evaluated = $this->evaluate($node, $params);
-
-        if ($evaluated instanceof $expectedType) {
-            return $evaluated;
-        }
-
-        throw new ExpressionError(sprintf(
-            'Expected "%s" but got "%s"', $expectedType, get_class($node)
-        ));
-    }
+    public function evaluateType(Node $node, string $expectedType, array $params): Node;
 
     /**
      * @param parameters $params
      */
-    public function evaluate(Node $node, array $params): Node
-    {
-        foreach ($this->evaluators as $evaluator) {
-            if (!$evaluator->evaluates($node)) {
-                continue;
-            }
-            $evaluated = $evaluator->evaluate($this, $node, $params);
-
-            return $evaluated;
-        }
-
-        throw new EvaluationError($node, sprintf(
-            'Could not find evaluator for node of type "%s"', get_class($node)
-        ));
-    }
+    public function evaluate(Node $node, array $params): Node;
 }
