+++ lib/Expression/NodeEvaluator/IntegerEvaluator.php
@@ -1,19 +1,20 @@
 <?php
 
-namespace PhpBench\Expression\Evaluator;
+namespace PhpBench\Expression\NodeEvaluator;
 
-use PhpBench\Expression\Ast\FloatNode;
+use PhpBench\Expression\Ast\IntegerNode;
 use PhpBench\Expression\Ast\Node;
 use PhpBench\Expression\Evaluator;
+use PhpBench\Expression\NodeEvaluator\AbstractEvaluator;
 
 /**
- * @extends AbstractEvaluator<FloatNode>
+ * @extends AbstractEvaluator<IntegerNode>
  */
-class FloatEvaluator extends AbstractEvaluator
+class IntegerEvaluator extends AbstractEvaluator
 {
     final public function __construct()
     {
-        parent::__construct(FloatNode::class);
+        parent::__construct(IntegerNode::class);
     }
 
     /**
