+++ lib/Extension/ExpressionExtension.php
@@ -11,20 +11,20 @@ use PhpBench\Expression\Ast\FunctionNode;
 use PhpBench\Expression\Ast\ParenthesisNode;
 use PhpBench\Expression\Ast\TolerableNode;
 use PhpBench\Expression\Evaluator;
-use PhpBench\Expression\Evaluator\ArgumentListEvaluator;
-use PhpBench\Expression\Evaluator\ArithmeticOperatorEvaluator;
-use PhpBench\Expression\Evaluator\BooleanEvaluator;
-use PhpBench\Expression\Evaluator\ComparisonEvaluator;
-use PhpBench\Expression\Evaluator\DisplayAsEvaluator;
-use PhpBench\Expression\Evaluator\FloatEvaluator;
-use PhpBench\Expression\Evaluator\FunctionEvaluator;
-use PhpBench\Expression\Evaluator\IntegerEvaluator;
-use PhpBench\Expression\Evaluator\ListEvaluator;
-use PhpBench\Expression\Evaluator\LogicalOperatorEvaluator;
-use PhpBench\Expression\Evaluator\ParameterEvaluator;
-use PhpBench\Expression\Evaluator\ParenthesisEvaluator;
-use PhpBench\Expression\Evaluator\TolerableEvaluator;
-use PhpBench\Expression\Evaluator\UnitEvaluator;
+use PhpBench\Expression\NodeEvaluator\ArgumentListEvaluator;
+use PhpBench\Expression\NodeEvaluator\ArithmeticOperatorEvaluator;
+use PhpBench\Expression\NodeEvaluator\BooleanEvaluator;
+use PhpBench\Expression\NodeEvaluator\ComparisonEvaluator;
+use PhpBench\Expression\NodeEvaluator\DisplayAsEvaluator;
+use PhpBench\Expression\NodeEvaluator\FloatEvaluator;
+use PhpBench\Expression\NodeEvaluator\FunctionEvaluator;
+use PhpBench\Expression\NodeEvaluator\IntegerEvaluator;
+use PhpBench\Expression\NodeEvaluator\ListEvaluator;
+use PhpBench\Expression\NodeEvaluator\LogicalOperatorEvaluator;
+use PhpBench\Expression\NodeEvaluator\ParameterEvaluator;
+use PhpBench\Expression\NodeEvaluator\ParenthesisEvaluator;
+use PhpBench\Expression\NodeEvaluator\TolerableEvaluator;
+use PhpBench\Expression\NodeEvaluator\UnitEvaluator;
 use PhpBench\Expression\ExpressionFunctions;
 use PhpBench\Expression\Func\MaxFunction;
 use PhpBench\Expression\Func\MeanFunction;
